requirements.txt
pbr
python-bugzilla
prettytable
ansible-runner
psutil
pathlib
jsonschema
PyYAML
ansible